requirements.txt
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0

pyg_lib~=0.2.0
torch_geometric~=2.3.0
torch_scatter~=2.1.0
torch_cluster~=1.6.1
torch_spline_conv~=1.2.2

pandas~=2.0.0
numpy~=1.23.5
scipy~=1.10.1
networkx~=2.8.4

wandb~=0.14.2
hydra-core~=1.3.2
